pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "file2md"
version = "0.1.0"
description = "Convert PDF and DOCX to clean, grep-friendly Markdown for AI/IDE workflows"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Faiz Khairi", email = "ifaizkhairi@gmail.com" }]
keywords = ["pdf", "docx", "markdown", "converter", "cli", "ai", "llm", "ide"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.1",
    "pymupdf>=1.23",
    "python-docx>=1.0",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "jinja2>=3.1",
    "python-multipart>=0.0.6",
]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
    "mypy>=1.0",
    "httpx>=0.24",
    "build>=1.0",
]
all = ["file2md[web,dev]"]

[project.scripts]
file2md = "file2md.cli:cli"

[project.urls]
Homepage = "https://github.com/faizkhairi/file2md"
Repository = "https://github.com/faizkhairi/file2md"
Issues = "https://github.com/faizkhairi/file2md/issues"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
